MULE-13046: In Db Module, it is necessary to use the scale for setting decimal values.
According to Oracle Documentation: for java.sql.Types.DECIMAL or java.sql.Types.NUMERIC types, it is necessary to indicate the number of digits after
the decimal point when we set statement parameters.
Now, we are ignoring this attribute for decimal values. Because of this, using some drivers, the decimal values are lost.