W-10673234: When a stored procedure is in a package Oracle returns duplicated parameter data if passed a null catalog. (#563)
When a stored procedure is in a package Oracle returns duplicated parameter data if passed a null catalog.

Connection's getCatalog() in Oracle always returns null anyway so catalogName here is always null. Now we pass "" as catalog if catalogName is null.

Other change: Close the previous metadata ResultSet if a new one will be requested.